+++ IncrementalInference/src/CliqueStateMachine/services/CliqStateMachineUtils.jl
@@ -350,7 +350,7 @@ Future
 """
 function approxCliqMarginalUp!(
   csmc::CliqStateMachineContainer,
-  childmsgs = LikelihoodMessage[];#fetchMsgsUpChildren(csmc, TreeBelief);
+  childmsgs = LikelihoodMessage[];
   N::Int = getCliqueSolverParams(csmc).N,
   dbg::Bool = getCliqueSolverParams(csmc).dbg,
   multiproc::Bool = getCliqueSolverParams(csmc).multiproc,
@@ -379,7 +379,7 @@ function approxCliqMarginalUp!(
         getWorkerPool(),
         fg_,
         cliqc,
-        csmc.solveKey,
+        csmc.csmoptions.solveKey,
         childmsgs,
         N,
         dbg,
@@ -401,10 +401,10 @@ function approxCliqMarginalUp!(
       @info "Single process upsolve clique=$(cliq.id)"
     end
     retdict =
-      upGibbsCliqueDensity(fg_, cliq, csmc.solveKey, childmsgs, N, dbg, iters, logger)
+      upGibbsCliqueDensity(fg_, cliq, csmc.csmoptions.solveKey, childmsgs, N, dbg, iters, logger)
 
     # DEBUG ON THE FLY
-    Bs = [HomotopyDensity_legacy(b) for (l,b) in retdict]
+    # Bs = [HomotopyDensity_legacy(b) for (l,b) in retdict]
     # @info "DX bw" string.(getBandwidth.(Bs))
 
   end